requirements.txt
annotated-types==0.7.0
bertopic==0.16.4
blis==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud==0.34.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
hdbscan==0.8.40
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pipdeptree==2.24.0
plotly==5.24.1
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
smart-open==7.0.5
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
